pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "pillow>=11.3.0",
    "playwright>=1.55.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
